update week
- the first alpha Form Application - see Releases
- fixed some structure mistakes

ATTENTION: the demo form application needs to be killed by hand in current time with the TaskManager.
Best testing is by starting GDB.EXE and shut down the gdb Thread from Task Manager - if you don't want to close and open again the developer Windows Console.

On my Desktop Development PC I use the FAR Manager - it is a Norton Commander Clone and it is Free to use under Windows.

Any feddback is welcome.
Drop a message on Questions

paule32